admin/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Django==3.0.4
django-crispy-forms==1.9.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pyTelegramBotAPI==3.6.7
pytest==5.4.2
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
Werkzeug==1.0.0
django-jsonfield==1.4.0
